Move message processing to its own module
